Move services to services directory
