+++ server/api/account.js
@@ -8,7 +8,7 @@ const util = require('../util.js');
 const nodemailer = require('nodemailer');
 const moment = require('moment');
 const monk = require('monk');
-const UserService = require('../repositories/UserService.js');
+const UserService = require('../services/UserService.js');
 
 let db = monk(config.dbPath);
 let userService = new UserService(db);
